Support for `Sidekiq.transactional_push!`
`Sidekiq.transactional_push!` is a feature that defers enqueuing Sidekiq workers inside a database transaction until after the after_commit callback. [sidekiq wiki](https://github.com/sidekiq/sidekiq/wiki/Advanced-Options#transactional-push)
When this feature was used in a multitenant environment, there was a bug where the tenant context switched within the transaction would be lost when enqueuing the job.

In this fix, it now preserves and carries over the multitenant context when enqueuing.